requirements.txt
# =============================================================================
# Reproducibility-pinned dependencies for
# "Foundation Model Embeddings Discard Small-Scale Signal in Chest Radiography"
#
# Tested with Python 3.10 on Ubuntu 22.04 + CUDA 12.1.
# Each version constraint is the minimum we have actually run; newer is fine.
# =============================================================================

# Core deep learning
torch>=2.1
transformers>=4.40
accelerate>=0.27
huggingface_hub>=0.20

# ML and stats
scikit-learn>=1.3
scipy>=1.11
statsmodels>=0.14
numpy>=1.24
pandas>=2.0
pyarrow>=14                        # parquet I/O

# Imaging
opencv-python-headless>=4.8
Pillow>=10
pydicom>=2.4                       # VinDr DICOM loading

# Notebook orchestration (papermill is the canonical runner)
papermill>=2.5
nbformat>=5.9
jupyter>=1.0
jupytext>=1.15                     # optional .py <-> .ipynb round-trip
nbstripout>=0.6                    # optional: keep diffs clean of execution outputs
tqdm>=4.65

# Visualisation (notebooks 05, 09, figure generation)
matplotlib>=3.7
umap-learn>=0.5                    # exp 05 falls back to PCA if absent
